backend/requirements.txt
# requirements.txt
numpy>=2.0.0,<3.0.0
pandas>=2.2.2,<3.0.0
markitdown[all]>=0.1.1
fastapi
uvicorn
sqlalchemy
alembic
python-jose[cryptography]
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
python-multipart
openai
google-cloud-vision
pdf2image
playwright
python-dotenv>=1.0.1
pydantic==2.9.1
pydantic-settings>=2.9.1,<3.0.0
holidays
celery
redis
pytest
httpx
psycopg2-binary
email-validator